fix warnings under postgres targets, re-enable `os::unix::{net,ucred}` as versions which return `unsupported`
