+++ Cargo.toml
@@ -60,7 +60,7 @@ object.debug = 0
 debug = 0
 [profile.dev.package]
 # Only use debuginfo=1 to further reduce compile times.
-bootstrap.debug = 1
+# bootstrap.debug = 1
