+++ library/std/src/os/net/mod.rs
@@ -1,5 +1,4 @@
 //! OS-specific networking functionality.
 
 #[cfg(any(target_os = "linux", target_os = "android", doc))]
-#[cfg(not(target_family = "postgres"))]
 pub(super) mod linux_ext;
