+++ library/std/src/os/unix/net/mod.rs
@@ -3,21 +3,9 @@
 #![allow(irrefutable_let_patterns)]
 #![stable(feature = "unix_socket", since = "1.10.0")]
 
-// TODO(thom): go through these modules again and use this in more places where
-// it currently does explicit `cfg`.
 #[cfg(target_family = "postgres")]
 #[allow(unused)]
-macro bail_if_postgres() {
-    // Make this conditional (even though the macro already is) just to avoid
-    // dead code warnings.
-    if cfg!(target_family = "postgres") {
-        return crate::sys::unsupported();
-    }
-}
-// `crate::sys::unsupported()` doesn't exist on non-postgres.
-#[cfg(not(target_family = "postgres"))]
-#[allow(unused)]
-macro bail_if_postgres() {}
+use super::bail_if_postgres;
 
 mod addr;
 #[doc(cfg(any(target_os = "android", target_os = "linux")))]
